requirements.txt
agate==1.7.1
annotated-types==0.7.0
asn1crypto==1.5.1
attrs==25.4.0
babel==2.17.0
backports.tarfile==1.2.0
boto3==1.42.23
botocore==1.42.23
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.0
dbt-core==1.7.19
dbt-extractor==0.5.1
dbt-semantic-interfaces==0.4.4
dbt-snowflake==1.7.1
filelock==3.20.2
fredapi==0.5.2
idna==3.11
importlib-metadata==6.11.0
isodate==0.6.1
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring==25.7.0
leather==0.4.1
Logbook==1.5.3
MarkupSafe==3.0.3
mashumaro==3.14
minimal-snowplow-tracker==0.0.2
more-itertools==10.8.0
msgpack==1.1.2
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pandas==2.3.3
parsedatetime==2.6
pathspec==0.11.2
platformdirs==4.5.1
protobuf==4.25.8
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
six==1.17.0
snowflake-connector-python==3.18.0
sortedcontainers==2.4.0
sqlparse==0.5.5
text-unidecode==1.3
tomlkit==0.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
zipp==3.23.0
